+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsetypescript.py
@@ -40,9 +40,9 @@ class EvaluatorResponseTypescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-03T05:48:17.067Z"
+    created: Optional[str] = "2026-07-03T08:32:24.507Z"
 
-    updated: Optional[str] = "2026-07-03T05:48:17.067Z"
+    updated: Optional[str] = "2026-07-03T08:32:24.507Z"
 
     updated_by_id: OptionalNullable[str] = UNSET