+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWK8B2EXDAXNXDJ0K8MSZCSM"
+        "01KWKHQSRS68S8ZN15JT3AYZ9M"
     )
     r"""The unique identifier of the data source"""