+++ packages/orq-rc/RELEASES.md
@@ -9239,3 +9239,13 @@ Based on:
 - [python v4.11.0-rc.74] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.74] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.74 - packages/orq-rc
+
+## 2026-07-03 08:34:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.788.0 (2.915.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.75] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.75] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.75 - packages/orq-rc